// client/src/app/shared/Models/basket.ts
import { v4 as uuidv4 } from 'uuid';

export interface IBasket{
  id: string;
  items: IBasketItem[]
}

export interface IBasketItem{
  id: number;
  gameName: string;
  price: number;
  quantity: number;
  pictureUrl: string;
  genre: string;
  device: string;
}

export class Basket implements IBasket{
  id = uuidv4();
  items : IBasketItem[] = [];

}

export interface IBasketTotals{
  shipping: number;
  subtotal: number;
  total: number
}